Push takeString logic to the lexers for better performance
The lexers can be more efficient about iterating through the bytes than
an external function. With takeString as a lexer method, there's no need
to perform multiple function calls for every byte. The overhead of the
function calls is very significant.

The byte lexer version is most efficient. It steps over the input
directly, and maintains a local variable with the current offset. When
it returns, it updates l.pos. Avoiding intermediate writes to l.pos
means the compiler can track the position in a register.

The reader lexer effectively inlines a simplified version of
peek()/take(). It avoids setting nextByte because this isn't used in the
context of parsing a string. It still has to call ReadByte() and
WriteByte(), though.

Benchmark results:

benchmark                      old ns/op     new ns/op     delta
BenchmarkUnmarshalMix          4389407       4472287       +1.89%
BenchmarkDecodeMix             1595274       1594484       -0.05%
BenchmarkBytesMix              1550321       1232586       -20.49%
BenchmarkReaderMix             3878020       3368036       -13.15%

The same optimization could potentially be applied for takeDigits(),
takeNumeric(), at the cost of more complexity in the lexer
implementations. The optimizations will probably be less helpful for
numeric types because strings can be much longer than numbers.